Moved md5Hash function to Tools
