+++ src/api/local/deployment.ts
@@ -1,5 +1,3 @@
-import Crypto from 'crypto';
-import { readFileSync } from 'fs';
 import createIgnore, { Ignore } from 'ignore';
 import path, { basename } from 'path';
 import tar from 'tar';
@@ -358,7 +356,7 @@ export namespace Deployment {
         const uploads: vscode.Uri[] = [];
         for await (const file of localFiles) {
           const remote = remoteMD5.find(e => e.path === file.path);
-          const md5 = md5Hash(file.uri);
+          const md5 = Tools.md5Hash(file.uri);
           if (!remote || remote.md5 !== md5) {
             uploads.push(file.uri);
           }
@@ -481,14 +479,6 @@ export namespace Deployment {
       md5: parts[0].trim(),
       path: parts[1].trim().substring(2) //these path starts with ./
     };
-  }
-
-  function md5Hash(file: vscode.Uri): string {
-    const bytes = readFileSync(file.fsPath);
-    return Crypto.createHash("md5")
-      .update(bytes)
-      .digest("hex")
-      .toLowerCase();
   }
 
   function toRelative(root: vscode.Uri, file: vscode.Uri) {
